Auto spell error log (fixes #1154) * When using auto spell, it will no longer print error messages to the map-server when no valid skill was selected
Note: This needs to be done because the client sends us random IDs if the user simply presses on "Ok" without selecting a skill, so it's technically not an error, but expected behavior that needs to be handled.

Special thanks to @anacondaqq and @stealth62 for helping to reproduce the error.