Unit command consistency changes
Made it so that all unit commands except for unitexists take 0 as the unit ID argument which will reference the activator of the script. Unitwarp has since previously had this functionality and it's very useful. Example use: When a monster walks over a npc script with the OnTouchNPC label unitkill 0; will kill the monster.