Cleanup and bug fix time
Fixed a bug introduced in bfe1355 where the client was not refreshed after collecting the achievement reward.
Fixed a racing condition if you tried to claim a reward right after completing the achievement.
Fixed that no RODEX mail will be sent if the achievement does not have any item reward.

Cleaned up the way the completion time and reward retrieval time are stored.
Also removed the redundant 'complete' flag.
The required database update script if you already had used the previous version will be supplied in the pull request.