Added support for 2018-01-24
In this client the new character select and cash shop UI from Ragnarok Zero were merged into the Sakray client.
Increased MAX_CHARS to 15 to fit the new character select
Added possibility to overwrite MAX_CHARS in defines_pre.h

As always thanks to @4144 for the shuffle packets and keys.