Add missing unit type documentation
The value `enable_baseatk` which has a default value set to 0x29F is represented in bit like this: `0010 1001 1111`.
The bit fields in the "Note 3" documentation only includes a description for the bit fields from 1 to 16. As we see from the code and from the default this field can affect more unit types which were added with this commit.