Cleaned up status_change_end
Removed leftover old debug info that made it necessary to the hide the real function behind a define.
Introduced a C++ default argument for timer id, as the timer id is only needed a single time, when the status expires naturally.
Replaced all currently existing calls that had INVALID_TIMER as argument and made use of the new default argument.