Removed player damage log (#8894)
It has been confirmed that players that log out stay in the damage log of the monsters, that they have attacked.
Once the monster dies their online status is being checked and players that are offline are ignored at that point.
This offline check is already implemented in rAthena additionally, since the player damage log seem to only have been used for MVPs.