Fixed SC_CRYSTALIZE
Crystalize had the NoCastCond flag in status.yml, but no source side condition was implemented.
Actually this was wrong and it is unconditional.
Also it has been confirmed that the movement restriction is also applying for monsters, so also the movement restriction is unconditional.

Additionally synced pre-renewal and renewal flags in status.yml

Thanks to @Playtester