Added a hasSCE function
This prevents to always call getSCE and have to check it for nullptr.
It should be used when it is only necessary to know, if a BL has a status active or not.
If you want to access values of a SCE you should still use getSCE and store the returned pointer.